requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
boto3==1.34.57
botocore==1.34.57
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contamination==1.0.0
contourpy==1.2.0
cycler==0.12.1
DataProperty==1.0.1
datasets==2.18.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docstring-parser==0.15
einops==0.7.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flash-attn==2.5.7
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==7.0.1
install==1.3.5
ipykernel==6.29.3
ipython==8.22.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.30.1
packaging==23.2
pandas==2.2.1
parso==0.8.3
pathvalidate==3.2.0
peft==0.9.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.42
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pybind11==2.11.1
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.18.0
s3transfer==0.10.0
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==68.2.2
shtab==1.7.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlitedict==2.1.0
stack-data==0.6.2
swagger-spec-validator==3.0.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
thefuzz==0.22.1
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
tqdm-multiprocess==0.0.11
traitlets==5.14.1
transformers==4.40.1
triton==2.2.0
trl==0.8.6
typepy==1.3.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tyro==0.7.3
tzdata==2024.1
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.13
webcolors==1.13
websocket-client==1.7.0
wheel==0.41.2
word2number==1.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
